LibraryApp/Tests/CategoriesControllerTests.cs
using LibraryApp.Controllers;
using LibraryApp.Models;
using Microsoft.AspNetCore.Mvc;
using Moq;
using Xunit;

public class CategoriesControllerTests
{
    [Fact]
    public void Index_ReturnsAViewResult_WithCategories()
    {
        /*
        // Arrange
        var mockCatRepo = new Mock<ICategoryRepository>();
        mockCatRepo.Setup(repo => repo.GetCategories())
                   .Returns(new List<Category>
                   {
                       new Category { Id = 1, Name = "Category 1" },
                       new Category { Id = 2, Name = "Category 2" },
                       new Category { Id = 3, Name = "Category 3" }
                   });

        var controller = new CategoriesController(mockCatRepo.Object);

        // Act
        var result = controller.Index();

        // Assert
        var viewResult = Assert.IsType<ViewResult>(result);
        var model = Assert.IsAssignableFrom<IEnumerable<Category>>(viewResult.Model);
        Assert.Equal(3, model.Count());*/
    }

    [Fact]
    public void Edit_WithValidId_ReturnsViewResult_WithCategory()
    {
        /*
        // Arrange
        int id = 1;
        var mockCatRepo = new Mock<ICategoryRepository>();
        mockCatRepo.Setup(repo => repo.GetCategoryById(id))
                   .Returns(new Category { Id = id, Name = "Category 1" });

        var controller = new CategoriesController(mockCatRepo.Object);

        // Act
        var result = controller.Edit(id);

        // Assert
        var viewResult = Assert.IsType<ViewResult>(result);
        var model = Assert.IsType<Category>(viewResult.Model);
        Assert.Equal(id, model.Id);
        */
    }

    [Fact]
    public void Edit_WithNullId_ReturnsViewResult_WithNewCategory()
    {
        /*
        // Arrange
        int? id = null;
        var controller = new CategoriesController();

        // Act
        var result = controller.Edit(id);

        // Assert
        var viewResult = Assert.IsType<ViewResult>(result);
        var model = Assert.IsType<Category>(viewResult.Model);
        Assert.Equal(0, model.Id); // Assuming default Id is 0 for new category
        */
    }
}
